% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Double.R
\name{plot.FuzzyDouble}
\alias{plot.FuzzyDouble}
\title{plot.FuzzyDouble}
\usage{
\method{plot}{FuzzyDouble}(x, xlab = "x", ylab = "Double", ...,
  LineCol = "red")
}
\arguments{
\item{LineCol}{}
}
\description{
plot.FuzzyDouble
}
